separate hierarchical outputs
